requirements.txt
# Requirements for circle detection scripts

# Traditional CV approach
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0

# Deep Learning approaches
# For YOLOv8
ultralytics>=8.0.0

# For DINOv3
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0
rasterio>=1.3.0

